a
a

+++ Principal/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render,redirect
 import requests
+import glob
 from datetime import date
 from Principal.forms import RegisterForm,LoginForm,NewProyecto
 from Principal.models import Usuario,Usuarios_proyecto,File,Proyecto
@@ -49,26 +50,25 @@ def Index(request):
     return render(request, 'Index.html')
 @login_required()
 def Proyectos(request):
-    form = NewProyecto(initial={'photo':'/default_image_project.png'})
+    form = NewProyecto(request.FILES,initial={'photo':'/default_image_project.png'},)
     UserInst = Usuario.objects.get(id=request.user.id)
     obj = Usuarios_proyecto.objects.filter(Usuario=request.user.id)
     Fileobj = File.objects.all();
     if request.method == "POST":
-        now = date.today()
-        Titulo = request.POST.get('Titulo')
-        if Titulo == None:
-            idprj = request.POST['idProyecto']
-            searchPrjct = Proyecto.objects.get(idProyecto=idprj)
-            clearListuser = Usuarios_proyecto.objects.get(Usuario=UserInst, Proyecto=searchPrjct)
-            clearListuser.delete()
-            searchPrjct.delete()
-        else:
-            Descripcion = request.POST.get('Descripcion')
-            photo = request.POST.get('photo')
-            SaveProject = Proyecto(Titulo=Titulo,Descripcion=Descripcion,Fecha_creacion=now,Usuario=UserInst,photo=photo)
-            SaveProject.save()
-            AddUserPrjt = Usuarios_proyecto(Usuario=UserInst,Proyecto=SaveProject)
-            AddUserPrjt.save()
+        if form.is_valid():
+            now = date.today()
+            Titulo = request.POST.get('Titulo')
+            if Titulo == None:
+                idprj = request.POST['idProyecto']
+                searchPrjct = Proyecto.objects.get(idProyecto=idprj)
+                clearListuser = Usuarios_proyecto.objects.get(Usuario=UserInst, Proyecto=searchPrjct)
+                clearListuser.delete()
+                searchPrjct.delete()
+            else:
+                Descripcion = request.POST.get('Descripcion')
+                photo = request.FILES.get('photo')
+                SaveProject = Proyecto(Titulo=Titulo,Descripcion=Descripcion,Fecha_creacion=now,Usuario=UserInst,photo=photo)
+                SaveProject.save()
 
     return render(request, 'Proyectos.html',{'Proyectos':obj,'form':form,'AllFiles':Fileobj})
 def Funciones(request):
